chore: remove libheif workaround (#598)
Someone on discord with previously affected hardware showed a
working bazzite kde the other day which never shipped this workaround.

We don't need this anymore, this must have gotten resolved upstream with 1.19.8
https://github.com/strukturag/libheif/releases/tag/v1.19.8

I tried an image with this change on my laptop that was previously affected
and this totally works now!

https://github.com/ublue-os/aurora/issues/8