feat: remove ugly discover symlink
causing all sorts of issues, one example:
https://github.com/ublue-os/aurora/issues/845

This will cause all of our users that installed Aurora before the
introduction of bazaar in Aug 2025 to have a missing/blank icon on the panel
with the name org.kde.discover.desktop.

New installations should not have this problem as their config in
plasma-org.kde.plasma.desktop-appletsrc points to the bazaar desktop file already.