feat: move back image specific files/hacks/tech debt (#1472)
* feat: move back rpm-ostreed.conf

This is tech debt only for this repo and is not an "asset" shared
between multiple aurora variants.

* feat: add bootc install config

CentOS Stream doesn't support BTRFS out of the box

* feat: add dracut/modprobe configs

module enablement is not an asset and should not be shared between
variants

* feat: add stuff for preinstall bazaar flatpak back

we don't need this on LTS as we start fresh there, this is not an "asset"

* feat: add SDDM stuff

Plasma Login Manager is the future and we shouldn't introduce
workarounds on other Aurora variants when there is a fix in sight.

* feat: add flatpak-nuke-fedora.service

We only have to deal with this problem on Fedora bases

* fix: actually COPY system_files

* Update system_files/shared/etc/rpm-ostreed.conf

Co-authored-by: Adam Fidel <ledif@users.noreply.github.com>

* chore: remove dracut zstd compression

* chore: move dracut config to command

---------

Co-authored-by: Adam Fidel <ledif@users.noreply.github.com>