fix: move starship init bash to /etc/profile.d
This should not be in common as LTS won't have starship on the image and
we can't get rid of it here.

For ZSH we are initting starship from /etc/zsh/zshrc from common.

We are currently not doing anything with Fish and therefore also not
initting starship. So Fish has the default prompt and no change in
behavior there.

Fixes: https://github.com/ublue-os/aurora/issues/1518